refactor(rust): split pipeline integration tests
Break the monolithic pipeline integration suite into focused test targets and move the shared harness into a dedicated pipeline support module.

Keep the failure-path cache tests fast by overriding retry settings so the split suite stays quick to run during iteration.
