+++ rust/tests/pipeline/mod.rs
@@ -1,0 +1,2 @@
+pub mod assertions;
+pub mod support;
